fix(createPlugin): change its signature to use full API object
I need to use the NR API object in order to call
`addCustomAttributes`, since custom attributes have extra bit of logic
unlike normal attributes. This means that instead of just passing
`newrelic.shim` into `createPlugin`, we need the actual `newrelic` API
object itself.

This is not a breaking change because this is all internal API.
Externally, in `index.js` we use `bind` to hide the first argument
passed into `createPlugin`, so users of this plugin should not see a
difference.