feat(createPlugin): add two configuration callbacks for custom attributes
Since we create two kinds of spans, there are two callbacks here, one
for each kind of span: operation and resolver spans.

The callbacks are passed the same arguments as the resolver and event
handlers so that they can use that information for building whatever
custom attributes they may want.