Send error extensions as custom attributes
The `GraphQLError`, which is used as a standard error type in
ApolloServer, defines a custom attribute of `extensions`[1] where
developers can add custom attributes to help them group errors. Send
through this attribute as custom attributes to NewRelic when catching
exceptions.

If users want to not pass through the extensions they can use the
standard attribute `include` and `exclude` rules to control them.

[1]: https://github.com/graphql/graphql-js/blob/688ee2f2a2f938e80bcf3808bc435d3d87a3ff3e/src/error/GraphQLError.ts